% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnuityMath.R
\name{Annuity_Payment}
\alias{Annuity_Payment}
\title{Annuity Payment Calculator
This function calculates the annuity payment whose present value matches a given PV at a given Discount Rate}
\usage{
Annuity_Payment(npv, discount_rate, number_years)
}
\arguments{
\item{npv}{The NPV of the annuity stream}

\item{discount_rate}{The rate for discounting the annuity stream}

\item{number_years}{How many years the annuity will be paid}
}
\description{
Annuity Payment Calculator
This function calculates the annuity payment whose present value matches a given PV at a given Discount Rate
}
\examples{
Annuity_Payment()
}
