% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2country.R
\name{coords2country}
\alias{coords2country}
\title{coords2country
This function returns the countries in which a set of points resides}
\usage{
coords2country(points)
}
\arguments{
\item{points}{a data.frame in which:column 1 contains the longitude in degrees and column 2 contains the latitude in degrees}
}
\description{
coords2country
This function returns the countries in which a set of points resides
}
\examples{
coods2country()
}
