% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2country.R
\name{country2ISO3}
\alias{country2ISO3}
\title{country2ISO3
This function returns the ISO3 abbreviation for a country}
\usage{
country2ISO3(country_name)
}
\arguments{
\item{country_name}{The country's name}
}
\description{
country2ISO3
This function returns the ISO3 abbreviation for a country
}
\examples{
country2ISO3()
}
