% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2country.R
\name{countrylist2ISO3}
\alias{countrylist2ISO3}
\title{countrylist2ISO3
This function returns the ISO3 abbreviation for a vector of countries}
\usage{
countrylist2ISO3(country_list)
}
\arguments{
\item{country_name}{The vector containing country names}
}
\description{
countrylist2ISO3
This function returns the ISO3 abbreviation for a vector of countries
}
\examples{
countrylist2ISO3()
}
