% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnuityMath.R
\name{DPBP}
\alias{DPBP}
\title{Discounted Payback Period
This function calculates the Discounted Payback Period of a stream of cashflows}
\usage{
DPBP(cashflows, npv, discount_rate)
}
\arguments{
\item{cashflows}{the cashflows to evaluate}

\item{npv}{the upfront cost of the investment}

\item{discount_rate}{the discount rate}
}
\description{
Discounted Payback Period
This function calculates the Discounted Payback Period of a stream of cashflows
}
\examples{
DPBP()
}
