% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnuityMath.R
\name{IRR}
\alias{IRR}
\title{IRR
This function calculates the Internal Rate of Return}
\usage{
IRR(cashflows, npv)
}
\arguments{
\item{cashflows}{the cashflows to evaluate}

\item{npv}{the upfront cost of the investment}
}
\description{
IRR
This function calculates the Internal Rate of Return
}
\examples{
IRR()
}
