% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnuityMath.R
\name{NPV}
\alias{NPV}
\title{NPV
This function calculates the Internal Rate of Return}
\usage{
NPV(cashflows, discount_rate)
}
\arguments{
\item{cashflows}{the cashflows to evaluate}

\item{npv}{the upfront cost of the investment}
}
\description{
NPV
This function calculates the Internal Rate of Return
}
\examples{
NPV()
}
