% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnuityMath.R
\name{PBP}
\alias{PBP}
\title{Payback Period (Non discounted)
This function calculates the Payback Period of a stream of cashflows}
\usage{
PBP(cashflows, NPV)
}
\arguments{
\item{cashflows}{the cashflows to evaluate}

\item{npv}{the upfront cost of the investment}
}
\description{
Payback Period (Non discounted)
This function calculates the Payback Period of a stream of cashflows
}
\examples{
PBP()
}
